windows: Support workload identity on GKE Windows node.
It first test if workload identity is enabled. If true, it downloads
gke-metadata-server.exe and starts it as native Windows service on
cbr0:988. Its metrics are exported on MGMT_IP:989 as gke-metrics-agent
scrape metrics on MGMT_IP.

Proxy policy will be applied to all pods on this node to redirect
169.254.169.254:80 to cbr0:988.
Access control list policy will be applied to cbr0 so that cbr0:988 can't be
accessed from outside the node.

Add a new kube_env ENABLE_WINDOWS_WORKLOAD_IDENTITY to control this
feature. After this PR is merged and cherry-picked, we can start to
allow creating Windows node with workload identity enabled and set it.

Ref: https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity
Bug: 154748130
Guitar tests bug: 159131141

Change-Id: I8a638336b5546b01acdd4e52d98528d919fa8a3d

Add logging agent config for gke-metadata-server to Support Workload Identity.

Ref: https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity
Bug: 161367090
Change-Id: I77d37f6e848dd58fcfc831d2a2ed4a0718cadd06

Fix fluent-bit configuration for Workload Identity logs.

The problem here was that WI fluent-bit configurations were
modeled for fluentd which has a different structure.

This is blocking fluent-bit support for GKE Windows, backporting to
GKE 1.20. This is an internal only change.

Bug: 183151146
Change-Id: Ie6ef71d8d23329492259a9b4c63f3ab6582770ce

Fix gke-internal fluent-bit for Windows defintions.

Based on PR: https://github.com/kubernetes/kubernetes/pull/101271
Bug: 166504674

Change-Id: I1772c66a606d10c37ca1142b07cadd61db11e6d1